Fixed crash according to DTS response
I opened a DTS incident and their response was very helpful:

In iOS 8 and above, implementors of UIViewControllerAnimatedTransitioning should retrieve the to and from views by sending a -viewForKey: message to the provided transition context, rather than sending a -viewControllerForKey: message and grabbing the view from the returned view controller.  This is because the navigation controller may provide a  fromView or toView that is different than the fromViewController.view or toViewController.view respectively.  In your particular example, the navigation controller actually wraps APLMainTableViewController's view and the presented Search Controller's view in an intermediate view and hands this intermediate view to the <UIViewControllerAnimatedTransitioning> as the fromView/toView (depending on the direction).  So if you're directly animating fromViewController.view or toViewController.view, you're animating the wrong views.